+++ CHANGES.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [2.0.0] - 2023-05-12
+
+### Changed
+
+- `Sidekiq::Pauzer.paused_queues` now returns names without `queue:` prefix.
+
+### Removed
+
+- `Sidekiq::Pauzer.paused_queue_names` was removed.
+
+
 ## [1.1.0] - 2023-05-12
 
 ### Deprecated
@@ -24,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release that supports sidekiq >= 6.5.0
 
 
-[unreleased]: https://gitlab.com/ixti/sidekiq-pauzer/-/compare/v1.0.0...main
+[unreleased]: https://gitlab.com/ixti/sidekiq-pauzer/-/compare/v2.0.0...main
+[2.0.0]: https://gitlab.com/ixti/sidekiq-pauzer/-/compare/v1.1.0...v2.0.0
 [1.1.0]: https://gitlab.com/ixti/sidekiq-pauzer/-/compare/v1.0.0...v1.1.0
 [1.0.0]: https://gitlab.com/ixti/sidekiq-pauzer/-/tree/v1.0.0